mainutil: Add function to reset signal handling
The use case I have here is an utility sub-command which reads from
stdin. With RootCommand's server-optimized signal handling, I'd have
to awkwardly watch for context cancellation for ctrl-C to work when
no input is delivered to the command of mine. Just restore SIGINT to
work normally for those cases.